requirements.txt
fastapi[all]==0.115.8
python-dotenv==1.0.1
sentry-sdk[fastapi]==2.20.0
asyncpg==0.30.0
pydantic[email]
firebase-admin